Refactor to use correct locale et_EE
