+++ src/Faker/Provider/et_EE/Address.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Faker\Provider\et_ET;
+namespace Faker\Provider\et_EE;
 
 class Address extends \Faker\Provider\Address
 {
