+++ src/Faker/Provider/et_EE/Person.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Faker\Provider\et_ET;
+namespace Faker\Provider\et_EE;
 
 class Person extends \Faker\Provider\Person
 {
